requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 'requirements\requirements.in'
#
aioredis==1.3.1           # via channels-redis
asgiref==3.3.1            # via channels, channels-redis, daphne, django
async-timeout==3.0.1      # via aioredis
attrs==20.3.0             # via automat, service-identity, twisted
autobahn==20.12.2         # via daphne
automat==20.2.0           # via twisted
cffi==1.14.4              # via cryptography
channels-redis==3.2.0     # via -r requirements\base-requirements.in
channels==3.0.2           # via -r requirements\base-requirements.in, channels-redis
constantly==15.1.0        # via twisted
cryptography==3.3.1       # via autobahn, pyopenssl, service-identity
daphne==3.0.1             # via channels
django-ipware==3.0.2      # via -r requirements\base-requirements.in
django-widget-tweaks==1.4.8  # via -r requirements\base-requirements.in
django==3.1.4             # via -r requirements\base-requirements.in, channels
hiredis==1.1.0            # via aioredis
hyperlink==20.0.1         # via twisted
idna==2.10                # via hyperlink, twisted
incremental==17.5.0       # via twisted
msgpack==1.0.1            # via channels-redis
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycparser==2.20           # via cffi
pyhamcrest==2.0.2         # via twisted
pyopenssl==20.0.1         # via twisted
pytz==2020.4              # via django
service-identity==18.1.0  # via twisted
six==1.15.0               # via automat, cryptography, pyopenssl
sqlparse==0.4.1           # via django
twisted[tls]==20.3.0      # via daphne
txaio==20.4.1             # via autobahn
zope.interface==5.2.0     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
